Finish up the explore page
