+++ lib/components/flight_card.dart
@@ -1,0 +1,154 @@
+import 'package:flutter/material.dart';
+import 'package:intl/intl.dart';
+
+class FlightCard extends StatelessWidget {
+  final String flightName, date, rating, discount;
+  final int realPrice, discountedPrice;
+  final formatCurrency = NumberFormat.simpleCurrency();
+
+  FlightCard(
+      {this.flightName,
+        this.date,
+        this.discount,
+        this.rating,
+        this.realPrice,
+        this.discountedPrice});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(top: 15.0, left: 5.0),
+      child: Stack(
+        children: <Widget>[
+          Padding(
+            padding: const EdgeInsets.only(right: 15.0),
+            child: Container(
+              height: 140.0,
+              decoration: BoxDecoration(
+                  borderRadius: BorderRadius.all(Radius.circular(10.0)),
+                  border: Border.all(
+                      color: Colors.grey,
+                      width: 1.0,
+                      style: BorderStyle.solid)),
+            ),
+          ),
+          Positioned(
+            top: 30.0,
+            left: 25.0,
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: <Widget>[
+                Row(
+                  children: <Widget>[
+                    Text(
+                      '${formatCurrency.format(discountedPrice)}',
+                      style: TextStyle(
+                        color: Colors.black,
+                        fontSize: 24.0,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                    SizedBox(
+                      width: 15.0,
+                    ),
+                    Text(
+                      '(${formatCurrency.format(realPrice)})',
+                      style: TextStyle(
+                        color: Colors.grey,
+                        fontSize: 18.0,
+                        fontWeight: FontWeight.normal,
+                        decoration: TextDecoration.lineThrough,
+                      ),
+                    ),
+                  ],
+                ),
+                SizedBox(
+                  height: 15.0,
+                ),
+                Padding(
+                  padding: const EdgeInsets.only(right: 10.0),
+                  child: Row(
+                    mainAxisSize: MainAxisSize.max,
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: <Widget>[
+                      CardChip(
+                        iconData: Icons.calendar_today,
+                        label: date,
+                      ),
+                      SizedBox(width: 5.0),
+                      CardChip(
+                        iconData: Icons.flight_takeoff,
+                        label: flightName,
+                      ),
+                      SizedBox(width: 5.0),
+                      CardChip(
+                        iconData: Icons.star,
+                        label: rating,
+                      )
+                    ],
+                  ),
+                )
+              ],
+            ),
+          ),
+          Positioned(
+            top: 7.0,
+            right: 5.0,
+            child: Container(
+              padding: EdgeInsets.symmetric(vertical: 5.0, horizontal: 10.0),
+              decoration: BoxDecoration(
+                color: Colors.orange[100],
+                borderRadius: BorderRadius.all(
+                  Radius.circular(10.0),
+                ),
+              ),
+              child: Text(
+                discount,
+                style: TextStyle(
+                  color: Theme.of(context).primaryColor,
+                  fontWeight: FontWeight.bold,
+                ),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class CardChip extends StatelessWidget {
+  final String label;
+  final IconData iconData;
+
+  CardChip({this.iconData, this.label});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: EdgeInsets.symmetric(vertical: 8.0, horizontal: 12.0),
+      decoration: BoxDecoration(
+        color: Colors.grey[200],
+        borderRadius: BorderRadius.all(
+          Radius.circular(15.0),
+        ),
+      ),
+      child: Row(
+        children: <Widget>[
+          Icon(
+            iconData,
+            size: 13.0,
+          ),
+          SizedBox(width: 10.0),
+          Text(
+            label,
+            style: TextStyle(
+              fontSize: 13.0,
+              fontWeight: FontWeight.normal,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
