--- lib/constants.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/material.dart';
-
-const dropDownItemsStyle = TextStyle(color: Colors.black, fontSize: 18.0);
-const dropDownMenuText = TextStyle(color: Colors.white, fontSize: 16.0);
-
-List<String> locations = [
-  'Boston (BS)',
-  'New York City (JFK)',
-  'Abuja (ABJ)',
-  'Lagos (LG)',
-  'Abu dhabi (AD)'
-];
-
-ThemeData appTheme = ThemeData(
-  primaryColor: Color(0xFFF3791A),
-  fontFamily: 'Oxygen'
-);
