+++ pubspec.yaml
@@ -30,6 +30,7 @@ dev_dependencies:
     sdk: flutter
   cloud_firestore: ^0.8.2
   cached_network_image:
+  rxdart:
 
 
 # For information on the generic Dart part of this file, see the
